API chanes

+++ app/src/emla/java/com/ramanbyte/emla/data_layer/repositories/ContentRepository.kt
@@ -8,7 +8,6 @@ import com.ramanbyte.data_layer.base.BaseRepository
 import com.ramanbyte.emla.data_layer.network.api_layer.SectionsController
 import com.ramanbyte.emla.models.ContentModel
 import com.ramanbyte.emla.models.MediaInfoModel
-import com.ramanbyte.utilities.AppLog
 import com.ramanbyte.utilities.FileUtils
 import com.ramanbyte.utilities.KEY_MEDIA_TYPE_AUDIO
 import com.ramanbyte.utilities.KEY_MEDIA_TYPE_VIDEO
@@ -133,7 +132,7 @@ class ContentRepository(mContext: Context) : BaseRepository(mContext) {
         } else {
         }*/
 
-        isMediaDownloaded = mediaInfoModel?.mediaStatus == Download.STATE_COMPLETED
+        isMediaDownloaded = mediaInfoModel?.mediaStatus.toInt() == Download.STATE_COMPLETED
 
         mediaInfoModel.apply {
             mimeType = FileUtils.getMimeType(contentLink) ?: ""
